trend and seasonality fully working for panel data
